t_bot/requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.8
pyTelegramBotAPI==3.6.7
python-dotenv==0.10.5
requests==2.22.0
urllib3==1.25.8
aiohttp==3.6.2
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
multidict==4.5.2
six==1.14.0
yarl==1.3.0